requirements.txt
Flask==0.10.1
Flask-Cors==2.1.2
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
Werkzeug==0.11.2
dotenv==0.0.5
itsdangerous==0.24
six==1.10.0
wsgiref==0.1.2
